refactor(suite-desktop): separate ethereum related tx summary components
